+++ src/revertJSONPatch.js
@@ -2,7 +2,8 @@ import { existsIn, getIn } from './immutabilityHelpers.js'
 import {
   immutableJSONPatch,
   isArrayItem,
-  preprocessJSONPatchOperation
+  parseFrom,
+  parsePath
 } from './immutableJSONPatch.js'
 import { compileJSONPointer } from './jsonPointer.js'
 import { startsWith } from './utils.js'
@@ -19,15 +20,25 @@ export function revertJSONPatch (json, operations, options) {
 
   immutableJSONPatch(json, operations, {
     before: (json, operation) => {
-      const revertOp = REVERT_OPS[operation.op]
-      if (!revertOp) {
-        return
+      let revertOperations
+      const path = parsePath(json, operation.path)
+      if (operation.op === 'add') {
+        revertOperations = revertAdd(json, path, operation.value)
+      } else if (operation.op === 'remove') {
+        revertOperations = revertRemove(json, path)
+      } else if (operation.op === 'replace') {
+        revertOperations = revertReplace(json, path)
+      } else if (operation.op === 'copy') {
+        revertOperations = revertCopy(json, path, operation.value)
+      } else if (operation.op === 'move') {
+        revertOperations = revertMove(json, path, parseFrom(operation.from))
+      } else if (operation.op === 'test') {
+        revertOperations = []
+      } else {
+        throw new Error('Unknown JSONPatch operation ' + JSON.stringify(operation.op))
       }
 
       let updatedJson
-      const preprocessedOperation = preprocessJSONPatchOperation(json, operation)
-      let revertOperations = revertOp(json, preprocessedOperation)
-
       if (options && options.before) {
         const res = options.before(json, operation, revertOperations)
         if (res && res.revertOperations) {
@@ -51,20 +62,12 @@ export function revertJSONPatch (json, operations, options) {
   return allRevertOperations
 }
 
-const REVERT_OPS = {
-  add: revertAdd,
-  remove: revertRemove,
-  replace: revertReplace,
-  copy: revertCopy,
-  move: revertMove
-}
-
 /**
  * @param {JSONData} json
- * @param {{ path: JSONPath }} operation
+ * @param {JSONPath} path
  * @return {JSONPatchDocument}
  */
-function revertReplace (json, { path }) {
+function revertReplace (json, path) {
   return [{
     op: 'replace',
     path: compileJSONPointer(path),
@@ -74,10 +77,10 @@ function revertReplace (json, { path }) {
 
 /**
  * @param {JSONData} json
- * @param {{ path: JSONPath }} operation
+ * @param {JSONPath} path
  * @return {JSONPatchDocument}
  */
-function revertRemove (json, { path }) {
+function revertRemove (json, path) {
   return [{
     op: 'add',
     path: compileJSONPointer(path),
@@ -87,35 +90,38 @@ function revertRemove (json, { path }) {
 
 /**
  * @param {JSONData} json
- * @param {{ path: JSONPath, value: JSONData }} operation
+ * @param {JSONPath} path
+ * @param {JSONData} value
  * @return {JSONPatchDocument}
  */
-function revertAdd (json, { path, value }) {
+function revertAdd (json, path, value) {
   if (isArrayItem(json, path) || !existsIn(json, path)) {
     return [{
       op: 'remove',
       path: compileJSONPointer(path)
     }]
   } else {
-    return revertReplace(json, { path, value })
+    return revertReplace(json, path, value)
   }
 }
 
 /**
  * @param {JSONData} json
- * @param {{ path: JSONPath, value: JSONData }} operation
+ * @param {JSONPath} path
+ * @param {JSONData} value
  * @return {JSONPatchDocument}
  */
-function revertCopy (json, { path, value }) {
-  return revertAdd(json, { path, value })
+function revertCopy (json, path, value) {
+  return revertAdd(json, path, value)
 }
 
 /**
  * @param {JSONData} json
- * @param {{ path: JSONPath, from: JSONPath }} operation
+ * @param {JSONPath} path
+ * @param {JSONPath} from
  * @return {JSONPatchDocument}
  */
-function revertMove (json, { path, from }) {
+function revertMove (json, path, from) {
   if (path.length < from.length && startsWith(from, path)) {
     // replacing the parent with the child
     return [
@@ -137,7 +143,7 @@ function revertMove (json, { path, from }) {
 
   if (!isArrayItem(json, path) && existsIn(json, path)) {
     // the move replaces an existing value in an object
-    revert = revert.concat(revertRemove(json, { path }))
+    revert = revert.concat(revertRemove(json, path))
   }
 
   return revert